requirements.txt
certifi==2021.5.30
chardet==4.0.0
Click==7.0
Flask==1.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2021.1
requests==2.25.1
six==1.16.0
twilio==6.59.1
urllib3==1.26.5
Werkzeug==0.16.0